requirements.txt
django
djangorestframework
djangorestframework-simplejwt
firebase-admin
django-cors-headers
pandas
openpyxl
prophet
psycopg2-binary
django-cors-headers
drf-yasg

